backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.7.0
python-multipart==0.0.17
